requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.4
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
filelock==3.6.0
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.1
